// web/src/api/models/ResponsesTasksResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ServicesTaskDTO } from './ServicesTaskDTO';
import {
    ServicesTaskDTOFromJSON,
    ServicesTaskDTOFromJSONTyped,
    ServicesTaskDTOToJSON,
    ServicesTaskDTOToJSONTyped,
} from './ServicesTaskDTO';

/**
 * 
 * @export
 * @interface ResponsesTasksResponse
 */
export interface ResponsesTasksResponse {
    /**
     * 
     * @type {Array<ServicesTaskDTO>}
     * @memberof ResponsesTasksResponse
     */
    data?: Array<ServicesTaskDTO>;
    /**
     * 
     * @type {string}
     * @memberof ResponsesTasksResponse
     */
    message?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ResponsesTasksResponse
     */
    success?: boolean;
}

/**
 * Check if a given object implements the ResponsesTasksResponse interface.
 */
export function instanceOfResponsesTasksResponse(value: object): value is ResponsesTasksResponse {
    return true;
}

export function ResponsesTasksResponseFromJSON(json: any): ResponsesTasksResponse {
    return ResponsesTasksResponseFromJSONTyped(json, false);
}

export function ResponsesTasksResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponsesTasksResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(ServicesTaskDTOFromJSON)),
        'message': json['message'] == null ? undefined : json['message'],
        'success': json['success'] == null ? undefined : json['success'],
    };
}

export function ResponsesTasksResponseToJSON(json: any): ResponsesTasksResponse {
    return ResponsesTasksResponseToJSONTyped(json, false);
}

export function ResponsesTasksResponseToJSONTyped(value?: ResponsesTasksResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(ServicesTaskDTOToJSON)),
        'message': value['message'],
        'success': value['success'],
    };
}