// web/src/api/models/ServicesTaskDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RepositoryTaskType } from './RepositoryTaskType';
import {
	RepositoryTaskTypeFromJSON,
	RepositoryTaskTypeFromJSONTyped,
	RepositoryTaskTypeToJSON,
	RepositoryTaskTypeToJSONTyped,
} from './RepositoryTaskType';

/**
 *
 * @export
 * @interface ServicesTaskDTO
 */
export interface ServicesTaskDTO {
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	completedAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	createdAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	description?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	dueAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	id?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	name?: string;
	/**
	 *
	 * @type {RepositoryTaskType}
	 * @memberof ServicesTaskDTO
	 */
	taskType?: RepositoryTaskType;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	taskTypeId?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	updatedAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ServicesTaskDTO
	 */
	userId?: string;
}

/**
 * Check if a given object implements the ServicesTaskDTO interface.
 */
export function instanceOfServicesTaskDTO(
	value: object,
): value is ServicesTaskDTO {
	return true;
}

export function ServicesTaskDTOFromJSON(json: any): ServicesTaskDTO {
	return ServicesTaskDTOFromJSONTyped(json, false);
}

export function ServicesTaskDTOFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean,
): ServicesTaskDTO {
	if (json == null) {
		return json;
	}
	return {
		completedAt:
			json['completed_at'] == null ? undefined : json['completed_at'],
		createdAt: json['created_at'] == null ? undefined : json['created_at'],
		description: json['description'] == null ? undefined : json['description'],
		dueAt: json['due_at'] == null ? undefined : json['due_at'],
		id: json['id'] == null ? undefined : json['id'],
		name: json['name'] == null ? undefined : json['name'],
		taskType:
			json['task_type'] == null
				? undefined
				: RepositoryTaskTypeFromJSON(json['task_type']),
		taskTypeId: json['task_type_id'] == null ? undefined : json['task_type_id'],
		updatedAt: json['updated_at'] == null ? undefined : json['updated_at'],
		userId: json['user_id'] == null ? undefined : json['user_id'],
	};
}

export function ServicesTaskDTOToJSON(json: any): ServicesTaskDTO {
	return ServicesTaskDTOToJSONTyped(json, false);
}

export function ServicesTaskDTOToJSONTyped(
	value?: ServicesTaskDTO | null,
	ignoreDiscriminator: boolean = false,
): any {
	if (value == null) {
		return value;
	}

	return {
		completed_at: value['completedAt'],
		created_at: value['createdAt'],
		description: value['description'],
		due_at: value['dueAt'],
		id: value['id'],
		name: value['name'],
		task_type: RepositoryTaskTypeToJSON(value['taskType']),
		task_type_id: value['taskTypeId'],
		updated_at: value['updatedAt'],
		user_id: value['userId'],
	};
}